# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GateAPI",
    "description": "LLM API aggregation gateway with model authenticity detection, multi-channel routing, and OpenAI/Anthropic compatible endpoints.",
    "version": "0.1.0",
    "license": { "name": "Apache-2.0" }
  },
  "servers": [{ "url": "/", "description": "Current server" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer" },
      "adminToken": { "type": "apiKey", "in": "header", "name": "x-admin-token" }
    },
    "schemas": {
      "ChatMessage": {
        "type": "object",
        "required": ["role", "content"],
        "properties": {
          "role": { "type": "string", "enum": ["system", "user", "assistant"] },
          "content": { "type": "string" }
        }
      },
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["model", "messages"],
        "properties": {
          "model": { "type": "string", "example": "gpt-4o" },
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/ChatMessage" } },
          "temperature": { "type": "number", "minimum": 0, "maximum": 2 },
          "top_p": { "type": "number" },
          "max_tokens": { "type": "integer" },
          "stream": { "type": "boolean", "default": false }
        }
      },
...
    }
  },
  "paths": {
    "/v1/chat/completions": {
      "post": {
        "tags": ["API"],
        "summary": "Chat Completions (OpenAI-compatible)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatCompletionRequest" } } } },
        "responses": {
          "200": { "description": "Completion response" },
          "401": { "description": "Invalid API key" },
          "404": { "description": "No channel for model" },
          "502": { "description": "All channels failed" }
        }
      }
    },
    "/v1/messages": {
      "post": {
        "tags": ["API"],
        "summary": "Messages (Anthropic-compatible)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Message response" } }
      }
    },
    "/v1/models": {
      "get": {
        "tags": ["API"],
        "summary": "List available models",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Model list" } }
      }
    },
    "/v1/embeddings": {
      "post": {
        "tags": ["API"],
        "summary": "Embeddings (passthrough)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Embedding response" } }
      }
    },
...
  }
}